[PropertyInfo] Cache Decorator
