+++ src/Symfony/Component/PropertyInfo/Tests/Extractors/SerializerExtractorTest.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\PropertyInfo\PropertyInfo\Tests\Extractors;
+namespace Symfony\Component\PropertyInfo\Tests\Extractors;
 
 use Doctrine\Common\Annotations\AnnotationReader;
 use Symfony\Component\PropertyInfo\Extractor\SerializerExtractor;
